Move SCNVector3 extensions to a separate file
